+++ client/src/components/BoxMain/BoxMain.css
@@ -1,0 +1,4 @@
+.screen {
+    width: 100%;
+    height: 85vh;
+  }
