+++ client/src/components/BoxMain/BoxMain.js
@@ -1,0 +1,3 @@
+export default {
+    name: "BoxMain",
+  };
