+++ client/src/components/ChatFooter/ChatFooter.js
@@ -1,0 +1,3 @@
+export default {
+    name: "ChatFooter",
+  };
