+++ client/src/components/ChatIntro/ChatIntro.css
@@ -1,0 +1,18 @@
+.intro {
+    height: 50vh;
+    width: 100%;
+  }
+  
+  .intro__container {
+    margin-left: 10%;
+    width: 40%;
+    color: white;
+  }
+  
+  h1 {
+    margin-top: 30px;
+    font-size: 50px;
+  }
+  .intro__container-text h4 {
+    font-size: 24px;
+  }
