+++ client/src/components/ChatIntro/ChatIntro.js
@@ -1,0 +1,1 @@
+export default {};
