+++ client/src/components/FormInput/FormInput.vue
@@ -13,44 +13,8 @@
   </div>
 </template> 
 
-<script>
-export default {
-  name: "FormInput",
-  props: {
-    title: {
-      type: String,
-    },
-    icon: {
-      type: String,
-    },
-    type: {
-      type: String,
-    },
-    info: {
-      type: String,
-    },
-    inputIndex: {
-      type: Number,
-    },
-  },
-  data() {
-    return {
-      inputContent: "",
-    };
-  },
+<script src="./FormInput.js">
 
-  methods: {
-    onInput() {
-      // create object with 2 properties to take value for changed input
-      let inputInfo = {
-        index: this.inputIndex,
-        content: this.inputContent,
-      };
-      // emit 'on-input' to parent
-      this.$emit(`on-input`, inputInfo);
-    },
-  },
-};
 </script>
 
 <style src="./FormInput.css" scoped>
