+++ client/src/components/SideBar/SideBar.js
@@ -1,0 +1,13 @@
+export default {
+	name: "SideBar",
+	data() {
+		return {
+			routes: [{ name: "Home", url: "/" }],
+		};
+	},
+	methods: {
+		onSignout() {
+			this.$router.push({ name: "Home" });
+		},
+	},
+};
