+++ client/src/screen/ChatLobby/ChatLobby.css
@@ -1,0 +1,5 @@
+.Chatlobby {
+	display: flex;
+	flex-direction: row;
+	align-content: space-between;
+}
