+++ client/src/screen/ChatLobby/ChatLobby.js
@@ -1,0 +1,10 @@
+import BoxChat from "../../components/BoxChat/BoxChat.vue";
+import SideBar from "../../components/SideBar/SideBar.vue";
+
+export default {
+  name: "ChatLobby",
+  components: {
+    SideBar,
+    BoxChat,
+  },
+};
