+++ client/src/screen/ChatLobby/ChatLobby.vue
@@ -5,23 +5,10 @@
   </div>
 </template>
 
-<script>
-import BoxChat from "../../components/BoxChat/BoxChat.vue";
-import SideBar from "../../components/SideBar/SideBar.vue";
+<script src="./ChatLobby.js">
 
-export default {
-  name: "ChatLobby",
-  components: {
-    SideBar,
-    BoxChat,
-  },
-};
 </script>
 
-<style>
-.Chatlobby {
-  display: flex;
-  flex-direction: row;
-  align-content: space-between;
-}
+<style src="./ChatLobby.css">
+
 </style>
