+++ client/src/screen/HomePage/HomePage.js
@@ -1,0 +1,11 @@
+import ChatHeader from "../../components/ChatHeaders/ChatHeader.vue";
+import ChatIntro from "../../components/ChatIntro/ChatIntro.vue";
+import ChatFooter from "../../components/ChatFooter/ChatFooter.vue";
+export default {
+	name: "HomePage",
+	components: {
+		ChatHeader,
+		ChatIntro,
+		ChatFooter,
+	},
+};
