+++ client/src/screen/HomePage/HomePage.vue
@@ -7,18 +7,8 @@
   </body>
 </template>
 
-<script>
-import ChatHeader from "../../components/ChatHeaders/ChatHeader.vue";
-import ChatIntro from "../../components/ChatIntro/ChatIntro.vue";
-import ChatFooter from "../../components/ChatFooter/ChatFooter.vue";
-export default {
-  name: "HomePage",
-  components: {
-    ChatHeader,
-    ChatIntro,
-    ChatFooter,
-  },
-};
+<script src="./HomePage.js">
+
 </script>
 
 <style scoped>
